Implementação para nao ter erros com o cors no front e ajustes de relacionamento nas entidades

+++ src/main/java/com/api/pa/controllers/RequestController.java
@@ -3,8 +3,9 @@ package com.api.pa.controllers;
 import com.api.pa.dtos.RequestDto;
 import com.api.pa.models.Product;
 import com.api.pa.models.Request;
+import com.api.pa.models.User;
 import com.api.pa.repositories.ProductRepository;
-import com.api.pa.services.ProductService;
+import com.api.pa.repositories.UserRepository;
 import com.api.pa.services.RequestService;
 import jakarta.validation.Valid;
 import org.springframework.beans.BeanUtils;
@@ -26,8 +27,7 @@ import java.util.Optional;
 public class RequestController {
 
     @Autowired
-
-    private ProductService productService;
+    private UserRepository userRepository;
     @Autowired
     private ProductRepository productRepository;
     final RequestService requestService;
@@ -41,21 +41,33 @@ public class RequestController {
         var requestModel = new Request();
         BeanUtils.copyProperties(requestDto, requestModel);
         Product product = productRepository.getReferenceById(requestDto.getProductId());
+        User user = userRepository.getReferenceById(requestDto.getUserId());
+        requestModel.setUser(user);
         requestModel.setProduct(product);
         requestModel.setDateIn(LocalDateTime.now(ZoneId.of("UTC")));
         return ResponseEntity.status(HttpStatus.OK).body(requestService.save(requestModel));
     }
+
     @GetMapping("requests")
-    public ResponseEntity<Page> getAllRequests (@PageableDefault(page = 0, size = 10, sort = "requestId", direction = Sort.Direction.ASC) Pageable pageable){
+    public ResponseEntity<Page> getAllRequests(@PageableDefault(page = 0, size = 10, sort = "requestId", direction = Sort.Direction.ASC) Pageable pageable) {
         return ResponseEntity.status(HttpStatus.OK).body(requestService.findAll(pageable));
     }
+
     @GetMapping("request/{requestId}")
-    public ResponseEntity<Object> getRequest (@PathVariable("requestId") Integer requestId){
+    public ResponseEntity<Object> getRequest(@PathVariable("requestId") Integer requestId) {
+        Optional<Request> requestOptional = requestService.findById(requestId);
+        if (!requestOptional.isPresent()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Request not found.");
+        }
         return ResponseEntity.status(HttpStatus.OK).body(requestService.findById(requestId));
     }
+
     @DeleteMapping("request/{requestId}")
-    public ResponseEntity<Object> deleteRequest (@PathVariable("requestId") Integer requestId){
+    public ResponseEntity<Object> deleteRequest(@PathVariable("requestId") Integer requestId) {
         Optional<Request> requestModelOptional = requestService.findById(requestId);
+        if(!requestModelOptional.isPresent()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Request not found.");
+        }
         requestService.delete(requestModelOptional.get());
         return ResponseEntity.status(HttpStatus.OK).body("Request deletado");
     }///
